+++ __init__.py
@@ -1,5 +1,5 @@
 """
 Oktopios — Un langage de programmation moderne et expressif 🐙
 """
-__version__ = "0.4.1"
+__version__ = "0.5.0"
 __author__ = "Mouanwiya Ali Soule"